Refactor IC hearing arrangements from comp/compForm to comp
